fixed folder structure and added usecase' interfaces
