+++ internal/pkg/event/usecase.go
@@ -1,1 +1,7 @@
 package event
+
+import "failless/internal/pkg/models"
+
+type UseCase interface {
+	InitEventsByTime(events []models.Event) (int, error)
+}
