+++ internal/pkg/tag/usecase.go
@@ -1,1 +1,7 @@
 package tag
+
+import "failless/internal/pkg/models"
+
+type UseCase interface {
+	InitEventsByTime(tags []models.Tag) (int, error)
+}
