+++ internal/pkg/user/usecase.go
@@ -1,6 +1,16 @@
 package user
 
-import "failless/internal/pkg/models"
+import (
+	"failless/internal/pkg/forms"
+	"failless/internal/pkg/models"
+)
+
+type UseCase interface {
+	UpdateUserInfo(form *forms.ProfileForm) (int, error)
+	GetUserInfo(profile *forms.ProfileForm) (int, error)
+	FillFormIfExist(cred *models.User) (int, error)
+	RegisterNewUser(user *forms.SignForm) error
+}
 
 // Get gender string by int id
 func GenderById(genderId int) string {
